added heat/nn/data_parallel.py and heat/nn/tests/test_data_parallel.py

+++ heat/nn/data_parallel.py
@@ -321,7 +321,7 @@ class DataParallel(tnn.Module):
                 (param_name, wait_handle, grad_loc.dtype, wrk)
             )
             # don't return grad_loc, otherwise gradient is doubled
-            return torch.zeros(*wrk.size(), dtype=grad_loc.dtype, device=grad_loc.device)
+            return torch.zeros(*wrk.size(), device=grad_loc.device)
 
         return _hook
